requirements.txt
streamlit
python-dotenv
groq
langchain-groq
langchain_community
langchain_core
langserve 
fastapi
uvicorn
